requirements.txt
matplotlib==3.3.0
plotly==4.9.0
pandas==1.3.5
numpy==1.19.5
lxml==4.5.2
html2text==2020.1.16
xlrd==1.2.0
openpyxl==3.0.5
altair==4.1.0
streamlit==1.1.0
cleantext==1.1.4
python_dateutil==2.8.2
